Move subscription related methods into new topics api module
