+++ src/lib/api/index.ts
@@ -5,3 +5,4 @@ export * as spaces from "./spaces";
 export * as resources from "./resources";
 export * as requests from "./requests";
 export * as access from "./access";
+export * as topics from "./topics";
