+++ src/routes/+layout.svelte
@@ -2,7 +2,7 @@
   import { init } from "$lib/channel";
   import { onMount } from "svelte";
   import Toasts from "../components/toasts.svelte";
-  import { calendars } from "$lib/api";
+  import { topics } from "$lib/api";
 
   onMount(() => {
     // Hacky workaround to only call "init" once in a Svelte HMR life-cycle.
@@ -13,7 +13,7 @@
     if (!("isInit" in window)) {
       init().then(() => {
         // After init subscribe to all calendars we know about.
-        calendars.subscribeToAll();
+        topics.subscribeToAll();
         
         // @TODO(sam): inform the backend of all authors we want to sync data with for each
         // calendar (add them to the topic map).
